+++ packages/angular-workspace/nimble-angular/CHANGELOG.md
@@ -1,8 +1,49 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Wed, 17 Jul 2024 21:19:45 GMT and should not be manually modified.
+This log was last generated on Mon, 22 Jul 2024 15:45:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 26.0.0
+
+Mon, 22 Jul 2024 15:45:21 GMT
+
+### Major changes
+
+- Update to Angular 17 ([ni/nimble@f69586a](https://github.com/ni/nimble/commit/f69586afcbdea5f4b8834d1123ec2b06d4eeecb3))
+- Bump @ni/nimble-components to v30.0.8
+
+## 25.3.3
+
+Fri, 19 Jul 2024 17:10:26 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.0.7
+
+## 25.3.2
+
+Fri, 19 Jul 2024 16:34:38 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.0.6
+
+## 25.3.1
+
+Fri, 19 Jul 2024 15:56:58 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.0.5
+
+## 25.3.0
+
+Thu, 18 Jul 2024 20:34:02 GMT
+
+### Minor changes
+
+- Export combobox page object from nimble-angular ([ni/nimble@ab27735](https://github.com/ni/nimble/commit/ab2773599bd4d2765f00666bcf5ff07c4936dfa6))
 
 ## 25.2.3