+++ packages/angular-workspace/spright-angular/CHANGELOG.md
@@ -1,8 +1,41 @@
 # Change Log - @ni/spright-angular
 
-This log was last generated on Wed, 17 Jul 2024 21:19:45 GMT and should not be manually modified.
+This log was last generated on Mon, 22 Jul 2024 15:45:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 3.0.0
+
+Mon, 22 Jul 2024 15:45:21 GMT
+
+### Major changes
+
+- Update to Angular 17 ([ni/nimble@f69586a](https://github.com/ni/nimble/commit/f69586afcbdea5f4b8834d1123ec2b06d4eeecb3))
+- Bump @ni/spright-components to v2.0.8
+
+## 2.0.7
+
+Fri, 19 Jul 2024 17:10:26 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v2.0.7
+
+## 2.0.6
+
+Fri, 19 Jul 2024 16:34:38 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v2.0.6
+
+## 2.0.5
+
+Fri, 19 Jul 2024 15:56:58 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v2.0.5
 
 ## 2.0.4